scripts/requirements.txt
pyserial
pyserial-asyncio
requests
python-dotenv